Wall Jump Fix
Character slides down walls without needing to be pressing buttons. Character can't move while in wall jump animation so that it can't climb a single wall. Character can't double jump from a wall. Added a tolerance timer for the character to jump right after starting to fall off a ledge or when turning against a wall.